requirements.txt
pytest==7.2.2
selenium==4.8.2
webdriver-manager==3.8.5
python-dotenv==1.0.0